tree-api/requirements/_base.txt
-e ../site-packages/py-smash/

aiohttp
aiohttp-swagger   # Swagger API Documentation builder for aiohttp server.
faker
treelib
click

fastjsonschema
pyyaml

